🚚 Rename: my-page/common 폴더 구조 ui, mutations로 분리
